Tohidul

+++ modules/blogs/blog.controller.js
@@ -30,8 +30,18 @@ exports.uploadsImage = async (req, res) => {
 
 // Create a blog
 exports.createBlog = async (req, res) => {
+  const {
+    category,
+    heroSection,
+    contentList,
+    Populer,
+    info,
+    tag,
+    thought,
+    learn,
+  } = req.body;
+
   try {
-    const { category, heroSection, contentList, Populer, info, tag } = req.body;
     const newBlog = new Blog({
       category,
       heroSection,
@@ -39,6 +49,8 @@ exports.createBlog = async (req, res) => {
       Populer,
       info,
       tag,
+      thought,
+      learn,
     });
 
     await newBlog.save();
@@ -47,6 +59,10 @@ exports.createBlog = async (req, res) => {
       blog: newBlog,
     });
   } catch (error) {
+    deleteImage(heroSection[0].headerImg);
+    contentList.map((content) => {
+      content.image && deleteImage(content.image);
+    });
     console.error(error);
     return res.status(500).json({
       message: "Error creating the blog post",
@@ -129,6 +145,7 @@ exports.updateSpecificFields = async (req, res) => {
 
 exports.updateContentFields = async (req, res) => {
   const { id, contentID } = req.params;
+  console.log(id, contentID);
   const { UpdatedImage, oldImage, headings, paragraphs } = req.body;
   const mainImage = UpdatedImage || oldImage;
 
@@ -141,6 +158,7 @@ exports.updateContentFields = async (req, res) => {
     paragraphs,
     image: mainImage,
   };
+  console.log(modifyObject);
 
   try {
     const blog = await Blog.findOne({ _id: id });
@@ -148,7 +166,7 @@ exports.updateContentFields = async (req, res) => {
     if (!blog) {
       return res.status(404).json({ message: "Blog not found" });
     }
-
+    console.log(blog);
     blog.contentList[contentID] = modifyObject;
     await blog.save();
 
@@ -164,7 +182,7 @@ exports.updateContentFields = async (req, res) => {
 exports.deleteContentAtIndex = async (req, res) => {
   try {
     const { id, contentID } = req.params; // Expect blogId and contentListIndex in URL params
-
+    console.log(id, contentID);
     const blog = await Blog.findById(id);
 
     if (!blog) {
@@ -173,6 +191,7 @@ exports.deleteContentAtIndex = async (req, res) => {
     if (contentID < 0 || contentID >= blog.contentList.length) {
       return res.status(400).json({ message: "Invalid index" });
     }
+    deleteImage(blog.contentList[contentID].image);
 
     blog.contentList.splice(contentID, 1);
     await blog.save();
@@ -210,9 +229,43 @@ exports.UpdateCategory = async (req, res) => {
     res.status(500).json({ error: error.message });
   }
 };
+exports.updateTexContent = async (req, res) => {
+  const { id } = req.params;
+  const { learn, thought } = req.body;
+  console.log(learn, thought);
+
+  try {
+    
+    const blog = await Blog.findById(req.params.id);
+    if (!blog) {
+      return res.status(404).json({ message: "Blog not found" });
+    }
+    const updatedBlog = await Blog.findOneAndUpdate(
+      { _id: id },
+      { learn, thought }, // Update object
+      { new: true }
+    );
+
+    if (!updatedBlog) {
+      return res.status(404).json({ error: "Blog not found" });
+    }
+
+    res.status(200).json(updatedBlog);
+  } catch (error) {
+    res.status(500).json({ error: error.message });
+  }
+};
 
 exports.deleteBlog = async (req, res) => {
   try {
+    const oneblog = await Blog.findById(req.params.id);
+    if (!oneblog) {
+      return res.status(404).json({ message: "Blog not found" });
+    }
+    deleteImage(oneblog.heroSection[0].headerImg);
+    oneblog.contentList.map((content) => {
+      content.image && deleteImage(content.image);
+    });
     const blog = await Blog.findByIdAndDelete(req.params.id);
     if (!blog) {
       return res.status(404).json({ message: "Blog not found" });
@@ -322,16 +375,15 @@ exports.getPopularBlogs = async (req, res) => {
 };
 exports.getBlogsGroupedByCategory = async (req, res) => {
   try {
-    
     const blogs = await Blog.find().lean();
 
     // Group blogs by category
     const groupedBlogs = blogs.reduce((result, blog) => {
-      const category = blog.category || "Uncategorized"; 
-      const heroSection = blog.heroSection[0] || ""; 
-      const blogDescription = blog.info || "No Description"; 
-      const createdAt = blog.createdAt; 
-      const tag = blog.tag || "No Tag"; 
+      const category = blog.category || "Uncategorized";
+      const heroSection = blog.heroSection[0] || "";
+      const blogDescription = blog.info || "No Description";
+      const createdAt = blog.createdAt;
+      const tag = blog.tag || "No Tag";
       id = blog._id;
 
       if (!result[category]) {
